Refactor routes into package modules
